Prevent newer instances of `Glicko2` from affecting existing `Player`s
The test I've added should help illustrate this bug.

Before this patch, the `defaultRating` and `volatility_algorithm` of the
`Player` class were defined upon the instantiation of a `Glicko2` instance.
Thus, new `Glicko2` instances would affect the `defaultRating` and
`volatility_algorithm` of already-instantiated `Player`s.

  var glicko = new glicko2.Glicko2({
    rating: 1400
  });
  var player = glicko.makePlayer();

  var newerGlicko = new glicko2.Glicko2({
    rating: 1600
  });

  player.getRating();
  // => 1600

As seen in the above code, `player` was created under `glicko`, which sets the
`defaultRating` to `1400`. However, the instantiation of `newerGlicko`
re-defines `Player.prototype.defaultRating` to `1600`, causing `player`'s rating
to change, despite having been created under `glicko`.

This fix moves the definition of `defaultRating` and `volatility_algorithm` to
`makePlayer()`. Instead of changing the `prototype` of `Player`, the `prototype`
of the specific `Player` created by `makePlayer()` is what's changed.

I've also fixed a couple of typos that I had made in #5.

Closes #6